requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.3.0
django-tinymce==3.6.1
bleach==6.1.0
python-decouple==3.8
django-filter==24.2
